LayoutTests/fullscreen/full-screen-request-rejected-expected.txt

EVENT(webkitfullscreenerror)
END OF TEST
This tests that a full-screen request which is not user-initiated will be rejected.
